added a dockerignore
Because of my symlinks, I need to ignore certain things so the
"setup" process can happen correctly when building the container.
I don't want to just "implicitly" have a theme that is already
installed be COPY'd into the image. Let the following step
do the appropriate copying into the image.